+++ resources/assets/js/main.js
@@ -32,4 +32,4 @@ window.Alpine.directive('prevent-submit-on-enter', preventSubmitOnEnter);
 window.Alpine.directive('copy', copyDirective);
 
 /* Tell Vite to build all files in this directory */
-import.meta.glob(['../img/**']);
+import.meta.glob(['../img/**', '../fonts/**']);