+++ resources/views/templates/page/_partials/favicon.blade.php
@@ -1,4 +1,4 @@
-<link rel="shortcut icon" href="{{ Vite::img('favicon/favicon.ico') }}" type="image/x-icon" />
+<link rel="shortcut icon" href="/chief/favicon/favicon.ico" type="image/x-icon" />
 <link rel="apple-touch-icon" sizes="180x180" href="{{ Vite::img('favicon/apple-touch-icon.png') }}" />
 <link rel="icon" type="image/png" sizes="32x32" href="{{ Vite::img('favicon/favicon-32x32.png') }}" />
 <link rel="icon" type="image/png" sizes="16x16" href="{{ Vite::img('favicon/favicon-16x16.png') }}" />