+++ src/Sites/UI/Livewire/WithLocaleToggle.php
@@ -43,6 +43,11 @@ trait WithLocaleToggle
     {
         $fields = Fields::make($components)->all();
 
+        // If the fragment allows fragments, we always show the locale toggle
+        if (isset($this->fragment) && $this->fragment->allowsFragments) {
+            return true;
+        }
+
         foreach ($fields as $field) {
             if ($field->showsLocales()) {
                 return true;