+++ src/Table/Livewire/Concerns/WithReordering.php
@@ -2,7 +2,7 @@
 
 namespace Thinktomorrow\Chief\Table\Livewire\Concerns;
 
-use Thinktomorrow\Chief\Shared\Concerns\Nestable\NestableTree;
+use Thinktomorrow\Chief\Shared\Concerns\Sortable\ReorderModels;
 
 trait WithReordering
 {
@@ -18,18 +18,42 @@ trait WithReordering
         $this->isReordering = false;
     }
 
-    public function getReorderResults(): NestableTree
+    public function getReorderResults(): iterable
     {
-        return NestableTree::fromIterable($this->getResults());
+        return $this->getResults();
     }
 
-    public function reorder()
+    public function reorder(array $orderedIds)
     {
-        // dd(func_get_args());
-        //        $reorderedPreviewFiles = collect($orderedIds)
-        //            ->map(fn ($previewFileId) => $this->previewFiles[$this->findPreviewFileIndex($previewFileId)])
-        //            ->all();
-        //
-        //        $this->previewFiles = $reorderedPreviewFiles;
+        if (count($orderedIds) < 1) {
+            return;
+        }
+
+        $this->verifyReorderingModelClass();
+
+        $modelClass = $this->getTable()->getReorderingModelClass();
+
+        app(ReorderModels::class)->handleByModel(new $modelClass, $orderedIds);
+    }
+
+    public function moveToParent($itemId, $parentId, array $orderedIds)
+    {
+        $this->verifyReorderingModelClass();
+
+        $modelClass = $this->getTable()->getReorderingModelClass();
+
+        // Get position of the item in the ordered list
+        $itemIndex = array_search($itemId, $orderedIds);
+
+        app(ReorderModels::class)->moveToParent(new $modelClass, $itemId, $parentId, $itemIndex);
+
+        $this->reorder($orderedIds);
+    }
+
+    private function verifyReorderingModelClass()
+    {
+        if (! $this->getTable()->hasValidReorderingModelClass()) {
+            throw new \RuntimeException('The table does not have a valid Sortable model class set. Given: ['.$this->getTable()->getReorderingModelClass().']');
+        }
     }
 }