+++ CHANGELOG.md
@@ -5,6 +5,19 @@ the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Unreleased
 
+This release introduces the new sorting UI for the table indices and contains breaking changes regarding the Sortable
+logic!
+
+- Fixed: show scopedLocale in edit parent fragment even when no localized fields are present so fragments can be
+  previewed per locale.
+- Added: **⚠ BC break** Sortable models now require the `Thinktomorrow\Chief\Shared\Concerns\Sortable\Sortable`
+  interface.
+- Changed: **⚠ BC break** Renamed `Thinktomorrow\Chief\Shared\Concerns\Sortable` trait to
+  `Thinktomorrow\Chief\Shared\Concerns\Sortable\SortableDefault`.
+- Changed: Actions::when() closure has now a first parameter of `component` instead of `model`, which is now second
+  parameter.
+- Changed: Make password 'missing login' error less clear for security reasons.
+
 ## [0.10.7] - 2025-06-03
 
 This release contains breaking changes!